pyproject.toml
[project]
name = "momem"
version = "0.1.0"
description = "CLI tool to manage reusable Python code snippets across projects"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.9",
]

[project.scripts]
momem = "momemcli.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["momemcli"]

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.coverage.run]
omit = ["momemcli/__main__.py"]
